requirements.txt
adjustText==0.7.3
articles==0.1.0
asgiref==3.5.0
colorama==0.4.4
cycler==0.11.0
Django==4.0.3
djangorestframework==3.13.1
fonttools==4.31.1
joblib==1.1.0
kiwisolver==1.4.0
many==0.6.9
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
patsy==0.5.2
Pillow==9.0.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
statsmodels==0.13.2
threadpoolctl==3.1.0
tqdm==4.63.0
tzdata==2022.1
